CodeFactor の警告を修正
