+++ tools/ToolBarTools/ToolBarImageMuxer/ToolBarImageMuxer.csproj
@@ -44,6 +44,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BmpFileComparer.cs" />
     <Compile Include="BmpMuxer.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
